+++ twitoff/models/User.py
@@ -7,6 +7,6 @@ class User(DB.Model):
         Database table for twitter users.
     """
 
-    id = DB.Column(DB.Integer, primary_key=True)
+    id = DB.Column(DB.BigInteger, primary_key=True)
     name = DB.Column(DB.String(50), nullable=False)
     username = DB.Column(DB.String(50), nullable=False)
